requirements.txt
python-telegram-bot==20.6
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.7
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
faiss-cpu>=1.8
sentence-transformers>=3.0
scikit-learn>=1.0
python-dotenv>=1.0
python-decouple==3.8
